backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
